+++ tests/testthat/test_Unary.R
@@ -1035,15 +1035,17 @@ test_that("cumprod_byname works as expected", {
 
 
 ###########################################################
-context("Replace NaN with 0")
+context("Replace NaN")
 ###########################################################
 
-test_that("replaceNaNWith0 works as expected", {
+test_that("replaceNaN works as expected", {
   expected <- matrix(c(1,0))
   suppressWarnings(a <- matrix(c(1, sqrt(-1))))
-  expect_equal(replaceNaNWith0(a), expected)
+  expect_equal(replaceNaN_byname(a), expected)
   # Should work with lists
-  expect_equal(replaceNaNWith0(list(a,a)), list(expected, expected))
+  expect_equal(replaceNaN_byname(list(a,a)), list(expected, expected))
+  # Try with a different value
+  expect_equal(replaceNaN_byname(a, 42), matrix(c(1,42)))
 })
 
 
@@ -1120,3 +1122,24 @@ test_that("count_vals_incols_byname works as expected", {
   expect_equal(count_vals_incols_byname(l, `>`, 2), list(ans, ans))
 })
 
+
+###########################################################
+context("Any")
+###########################################################
+
+test_that("any_byname works as expected", {
+  m <- matrix(rep(TRUE, times = 4), nrow = 2, ncol = 2)
+  expect_true(all_byname(m))
+  expect_true(any_byname(m))
+  
+  n <- matrix(c(TRUE, FALSE), nrow = 2, ncol = 1)
+  expect_false(all_byname(n))
+  expect_true(any_byname(n))
+  
+  # Also works for lists
+  expect_equal(all_byname(list(m,m)), list(TRUE, TRUE))
+  expect_equal(any_byname(list(m,m)), list(TRUE, TRUE))
+  expect_equal(all_byname(list(n,n)), list(FALSE, FALSE))
+  expect_equal(any_byname(list(n,n)), list(TRUE, TRUE))
+  
+})