+++ NEWS.md
@@ -1,5 +1,11 @@
 # News for `matsbyname`
 
+## matsbyname 0.2.8 (2018-05-17)
+
+* New functions `all_byname` and `any_byname` make logical tests easy.
+* New function `replaceNaN_byname` replaces `NaN` entries with a value (default is 0).
+
+
 ## matsbyname 0.2.7 (2018-04-15)
 
 * Refactored most `*col*_byname` functions to call their respective `*row*_byname` functions